requirements.txt
google-genai>=1.0.0
PyMuPDF>=1.24.0
